Update arduino program -discreteintensity
Analog reading of the photocell is taken through the sensor . If the photocell reading is less than 200 , it is dark in the room and led1 is glows. If the photocell reading is less than 400 , it is dim and led2 is glows. If the photocell reading is less than 500 , it is light and led3 is glows.If the photocell reading is less than 800, it is bright and led2 is glows otherwise it is very bright and no light glows.